requirements.txt
azure-mgmt-resource
azure-mgmt-storage
azure-identity
azure-mgmt-eventhub